Change setup.py to split_init.py
Renamed setup.py to split_init.py to reserve
setup.py name for future use

Work in progress https://github.com/crisbal/Album-Splitter/issues/17
